moves stateful components into containers folder
