+++ src/index.js
@@ -1,6 +1,6 @@
 import './index.css';
 
-import App from './app/App';
+import App from './App';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import React from 'react';
