+++ src/pages/signInSignUp/SignInSignUp.jsx
@@ -1,8 +1,8 @@
 import './SignInSignUp.styles.scss';
 
 import React from 'react';
-import SignIn from 'components/signIn';
-import SignUp from 'components/signUp';
+import SignIn from 'containers/signIn';
+import SignUp from 'containers/signUp';
 
 const SignInSignUp = props => (
 	<div className='sign-in-and-sign-up'>
